Add eslint
