+++ src/views/Home.jsx
@@ -1,19 +1,16 @@
 import React, {useEffect} from 'react';
-import {Button, Container, Col, Image, Row, Ratio} from 'react-bootstrap';
+import {Button, Card, Container, Col, Image, Row} from 'react-bootstrap';
 
-import Johann710Banner from '../assets/images/VRChat_1920x1080_2020-04-05_13-18-23.897.png';
+import ScrollReveal from 'scrollreveal';
+
 import LilyPFP from '../assets/images/LilyPFP.png';
-import LilyBanner from '../assets/images/VRChat_1920x1080_2020-11-11_19-48-16.201.png';
-import SourBanner from '../assets/images/VRChat_1920x1080_2020-11-11_22-56-07.930.png';
 
-const Home = () => {
+const HomeNew = () => {
   useEffect(() => {
     document.title = `KyButler's Site - Home`;
-  });
 
-  const onClickButton = (link, name) => {
-    window.open(link, '_blank');
-  };
+    ScrollReveal().reveal('.sr', {delay: 100});
+  });
 
   const buttons = [
     {
@@ -51,121 +48,151 @@ const Home = () => {
       link: 'https://trello.com/b/oG4w3rhH/twitchtv-kybutler',
       colorName: 'trello',
     },
+    {
+      title: 'Tiktok',
+      link: 'https://www.tiktok.com/@kybutler_?',
+      colorName: 'tiktok',
+    },
   ];
 
   return (
     <>
-      <Container style={{marginBottom: '3em', marginTop: '1em'}}>
+      <Container className="sr">
         <Row>
-          <Col xs={{order: 'last'}} md={{order: 'first'}} xs={12}>
-            <div>
-              <h1>
-                Hi, I'm <span className="kybutler-text">KyButler!</span>
-              </h1>
-
-              <p>
-                . . . a recently graduated university student who has too much time for videogames - hopefully you'll enjoy watching me play some as
-                much as I like playing them!
-              </p>
-            </div>
+          <Col className="mt-4" xs={{order: 'last', span: 12}} md={{order: 'first', span: 6}}>
+            <Card className="shadow">
+              <Card.Body>
+                <Card.Title>
+                  <p className="h1">
+                    Hi, I&apos;m
+                    <span className="kybutler-text"> KyButler! </span>
+                  </p>
+                </Card.Title>
+                <Card.Text>
+                  . . . a full-time software developer with a passion for playing videogames, hopefully you&apos;ll enjoy watching me play some as
+                  much as I enjoy playing them!
+                </Card.Text>
+              </Card.Body>
+            </Card>
           </Col>
-          <Col xs={{order: 'first'}} md={{order: 'last'}}>
-            <Image src={LilyPFP} style={{marginBottom: '1em', borderRadius: '5px'}} fluid />
+          <Col className="mt-4" xs={{order: 'first'}} md={{order: 'last'}}>
+            <Image className="shadow" fluid src={LilyPFP} rounded />
           </Col>
         </Row>
+      </Container>
 
-        {buttons.map((button, i) => (
-          <Row key={i}>
-            <Col>
-              <div className="button-container">
-                <Button
-                  onClick={() => onClickButton(button.link, button.colorName)}
-                  className={`mb-2 ${button.colorName}-color`}
-                  style={{color: 'white', cursor: 'pointer', height: '3em'}}
-                  variant="dark"
-                  block>
-                  <center>{button.title}</center>
-                </Button>
-              </div>
-            </Col>
-          </Row>
+      <Container className="sr">
+        {buttons.map(button => (
+          <Button
+            className={`mt-4 shadow sr ${button.colorName}-color`}
+            key={button.title}
+            onClick={() => window.open(button.link, '_blank')}
+            size={'lg'}
+            style={{color: 'white', width: '100%'}}
+            variant="outline-dark">
+            {button.title}
+          </Button>
         ))}
       </Container>
 
-      <Container style={{marginBottom: '3em'}}>
-        <hr />
+      <Container className="sr">
+        <hr className="mt-4" />
+      </Container>
+
+      <Container className="sr">
         <Row>
-          <Col xs={{order: 'last'}} md={{order: 'first'}} xs={12}>
-            <Ratio aspectRatio="1x1">
-              <iframe
-                style={{
-                  border: 0,
-                  width: '100%',
-                  height: '100%',
-                  borderRadius: '5px',
-                }}
-                src="https://bandcamp.com/EmbeddedPlayer/track=2833081214/size=large/bgcol=333333/linkcol=fe7eaf/tracklist=false/transparent=true/"
-                seamless>
-                <a href="https://thesourknights.bandcamp.com/track/the-kybutler-collective">The KyButler Collective by The Sour Knights</a>
-              </iframe>
-            </Ratio>
+          <Col className="mt-4" xs={{order: 'last', span: 12}} md={{order: 'first', span: 6}}>
+            <iframe
+              className="shadow"
+              style={{border: 0, width: '100%', height: '120px'}}
+              src="https://bandcamp.com/EmbeddedPlayer/track=2833081214/size=large/bgcol=333333/linkcol=fe7eaf/tracklist=false/artwork=small/transparent=true/"
+              seamless>
+              <a href="https://thesourknights.bandcamp.com/track/the-kybutler-collective">The KyButler Collective by The Sour Knights</a>
+            </iframe>
           </Col>
-          <Col xs={{order: 'first'}} md={{order: 'last'}}>
-            <div>
-              <h1>
-                The <span className="kybutler-text">KyButler</span> Collective
-              </h1>
-              <hr />
-              <p>
-                This is a track created by{' '}
-                <span className="sour-text" onClick={() => window.open('https://www.youtube.com/channel/UC1YhDvjc50ax8D_RlwdLB6Q', '_blank')}>
-                  The Sour Knights
-                </span>{' '}
-                dedicated to, well . . . me! I love this song and I love the work that Sour puts into his music, definitely worth checking out!
-              </p>
-            </div>
+          <Col className="mt-4" xs={{order: 'first'}} md={{order: 'last'}}>
+            <Card className="shadow">
+              <Card.Body>
+                <Card.Title>
+                  <p className="h1">
+                    The
+                    <span className="kybutler-text"> KyButler </span>
+                    Collective
+                  </p>
+                </Card.Title>
+                <Card.Text>
+                  This is a track created by <span className="sour-text">The Sour Knights</span> dedicated to, well . . . me! I love this song and I
+                  love the work that <span className="sour-text">Sour</span> puts into his music, definitely worth checking out!
+                </Card.Text>
+              </Card.Body>
+            </Card>
           </Col>
         </Row>
       </Container>
 
-      <Container style={{marginBottom: '10em'}}>
-        <hr />
+      <Container className="sr">
+        <hr className="mt-4" />
+      </Container>
+
+      <Container className="sr">
         <Row>
-          <Col>
-            <Image
-              className="image-on-hover"
-              onClick={() => window.open('https://www.twitch.tv/andynya', '_blank')}
-              style={{borderRadius: '5px', marginBottom: '1em'}}
-              src={LilyBanner}
-              fluid
-            />
+          <Col className="mt-4" xs={{order: 'first', span: 12}} md={{order: 'last', span: 6}}>
+            <iframe
+              className="shadow"
+              src="https://discord.com/widget?id=571371327177621504&theme=dark"
+              width="350"
+              height="500"
+              allowtransparency="true"
+              frameBorder="0"
+              sandbox="allow-popups allow-popups-to-escape-sandbox allow-same-origin allow-scripts"></iframe>
+          </Col>
+          <Col className="mt-4" xs={{order: 'last'}} md={{order: 'first'}}>
+            <Card className="shadow">
+              <Card.Body>
+                <Card.Title>
+                  <p className="h1">
+                    The
+                    <span className="discord-text"> Discord </span>
+                    Server
+                  </p>
+                </Card.Title>
+                <Card.Text>
+                  We have a community Discord server with over 500 people in it! Stop by, say hi, and have a great time meeting people, sharing memes,
+                  and more!
+                </Card.Text>
+              </Card.Body>
+            </Card>
           </Col>
         </Row>
+      </Container>
+
+      <Container className="sr">
+        <hr className="mt-4" />
+      </Container>
+
+      <Container className="sr">
         <Row>
-          <Col>
-            <Image
-              className="image-on-hover"
-              onClick={() => window.open('https://www.twitch.tv/j0hann710', '_blank')}
-              style={{borderRadius: '5px', marginBottom: '1em'}}
-              src={Johann710Banner}
-              fluid
-            />
+          <Col className="mt-4">
+            <Card className="shadow">
+              <Card.Body>
+                <Card.Title>
+                  <p className="h1">So what are you waiting for?</p>
+                </Card.Title>
+                <Card.Text>
+                  Follow some of my socials, join up on my streams, say hi in the Discord, and most importantly, enjoy your time as part of the
+                  KyButler community! I look forward to meeting you! - KyButler
+                </Card.Text>
+              </Card.Body>
+            </Card>
           </Col>
         </Row>
-        <Row>
-          <Col>
-            <Image
-              className="image-on-hover"
-              onClick={() => window.open('https://www.youtube.com/channel/UC1YhDvjc50ax8D_RlwdLB6Q', '_blank')}
-              style={{borderRadius: '5px', marginBottom: '1em'}}
-              src={SourBanner}
-              fluid
-            />
-          </Col>
-        </Row>
+      </Container>
+
+      <Container className="sr">
+        <hr className="mt-4" />
       </Container>
     </>
   );
 };
 
-export default Home;
+export default HomeNew;
