+++ src/views/Trello.jsx
@@ -5,7 +5,7 @@ const Trello = () => {
     document.title = `KyButler's Site - Trello`;
   });
 
-  return <iframe src="https://trello.com/b/oG4w3rhH.html" frameBorder="0" width="100%" height="100%" onLoad={() => setLoading(false)}></iframe>;
+  return <iframe src="https://trello.com/b/oG4w3rhH.html" frameBorder="0" width="100%" height="100%"></iframe>;
 };
 
 export default Trello;
